pe-org-air-platform/pyproject.toml
[project]
name = "pe-orgair-platform"
version = "0.1.0"
description = ""
authors = [
    {name = "AakashBelide",email = "aakash.belide@gmail.com"}
]
readme = "README.md"
packages = [{ include = "pe_orgair", from = "src" }]
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.128.0,<0.129.0)",
    "uvicorn[standard] (>=0.40.0,<0.41.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "snowflake-connector-python (>=3.10.0,<4.0.0)",
    "sqlalchemy (>=2.0.45,<3.0.0)",
    "snowflake-sqlalchemy (>=1.6.1,<2.0.0)",
    "alembic (>=1.18.1,<2.0.0)",
    "redis (>=7.1.0,<8.0.0)",
    "structlog (>=25.5.0,<26.0.0)",
    "sse-starlette (>=3.2.0,<4.0.0)",
    "websockets (>=16.0,<17.0)",
    "boto3 (>=1.42.30,<2.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "tenacity (>=9.1.3,<10.0.0)",
    "sec-edgar-downloader (>=5.1.0,<6.0.0)",
    "aiohttp (>=3.9.3,<4.0.0)",
    "playwright (>=1.41.2,<2.0.0)",
    "playwright-stealth (>=1.0.6,<2.0.0)",
    "python-jobspy (>=1.1.48,<2.0.0)",
    "pandas (>=2.0.0,<3.0.0)",
    "pdfkit (>=1.0.0,<2.0.0)",
    "beautifulsoup4 (>=4.12.0,<5.0.0)",
    "pdfplumber (>=0.11.0,<0.12.0)",
    "lxml (>=6.0.0,<7.0.0)",
    "pyrate-limiter (>=3.9.0,<4.0.0)"
]



[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^9.0.2"
pytest-asyncio = "^1.3.0"
pytest-cov = "^7.0.0"
black = "^25.12.0"
ruff = "^0.14.13"
mypy = "^1.19.1"
hypothesis = "^6.150.2"
factory-boy = "^3.3.3"

